+++ rollup.config.mjs
@@ -1,4 +1,3 @@
-import json from "@rollup/plugin-json";
 import terser from "@rollup/plugin-terser";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
@@ -19,6 +18,5 @@ export default {
             moduleDirectories: ["node_modules"],
         }),
         commonjs(),
-        json(),
     ],
 };