+++ index.js
@@ -130,10 +130,10 @@ remarklet.deactivate = function () {
 
 /**
  * Get the version number of the Remarklet library in use.
- * @type {"1.2.12"}
+ * @type {"1.2.13"}
  * @readonly
  * @since 1.0.2
  */
-remarklet.version = "1.2.12";
+remarklet.version = "1.2.13";
 
 export default remarklet;